Split hook handler…
…as LoadExtensionSchemaUpdatesHook does not allow has a service
dependency
